refactor(benchmarks): drop redundant float32 cast in _ObsChannelFilterEnv._filter_obs
Code-review follow-up on PR-B. The inner MPECooperativePushEnv emits
float32 state already; the np.asarray with dtype=float32 narrows it
defensively against a future upstream wrapper, and the slice
``state[indices]`` carries the dtype unchanged. The trailing
``.astype(np.float32)`` was a no-op cast that allocated a copy.

Behavior unchanged; all 17 OM tier-1 tests still pass.

+++ src/chamber/benchmarks/stage1_om.py
@@ -243,8 +243,11 @@ class _ObsChannelFilterEnv(gym.Env):  # type: ignore[type-arg]
         indices = list(self._channel_indices)
         filtered_agents: dict[str, dict[str, NDArray[np.float32]]] = {}
         for uid, agent_obs in obs["agent"].items():
+            # The inner MPE env emits float32 state already; the
+            # ``np.asarray`` is defensive (in case a future wrapper
+            # upstream returns a list) and the slice carries dtype.
             state = np.asarray(agent_obs["state"], dtype=np.float32)
-            filtered_agents[uid] = {"state": state[indices].astype(np.float32)}
+            filtered_agents[uid] = {"state": state[indices]}
         return {"agent": filtered_agents}
